# docs/swagger.yaml
basePath: /financial_tracker
definitions:
  api.ErrorResponse:
    properties:
      error: {}
    type: object
  api.SuccessResponse:
    properties:
      value: {}
    type: object
  categoryHandlers.RequestCreateCategory:
    properties:
      description:
        example: Shopping in the store
        type: string
      limit:
        type: integer
      name:
        example: jonn
        type: string
    required:
    - description
    - limit
    - name
    type: object
  categoryHandlers.RequestUpdateCategory:
    properties:
      category_id:
        example: 2
        type: integer
      description:
        example: car expenses
        type: string
      limit:
        example: 2000
        type: integer
      name:
        example: jonn
        type: string
    required:
    - description
    - limit
    - name
    type: object
...
host: localhost:8080
info:
  contact:
    email: asssv0423348@gmail.com
    name: Ярослав
    url: https://github.com/goggle-source
  description: API для работы с финансовым трекером
  license:
    name: MIT
  title: Финансовый Трекер
  version: "1.0"
paths:
  /category/create:
    post:
      consumes:
      - application/json
      description: Создание новой категории для зарегистрировшегося пользователя
      parameters:
      - description: данные для создание новой категории
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/categoryHandlers.RequestCreateCategory'
      produces:
      - application/json
      responses:
        "200":
          description: Успешное создание категории
          schema:
            $ref: '#/definitions/api.SuccessResponse'
        "400":
          description: Некорректный данные
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Ошибка авторизации
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "501":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - jwtAuth: []
      summary: Создание Категории
      tags:
      - categories
  /category/delete/{id}:
    delete:
      description: Удаление категории конкретного пользователя
      parameters:
      - description: userID для удаление категории
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/api.SuccessResponse'
        "400":
          description: Некорректные данные
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Ошибка авторизации
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Категория не найдена
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - jwtAuth: []
      summary: Удаление категории
      tags:
      - categories
  /category/get//{id}:
    get:
      description: Получение категории для конкретного пользователя
      parameters:
      - description: userID для получение категории
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/api.SuccessResponse'
        "400":
          description: Некорректный данные
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Ошибка авторизации
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Категория не найдена
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - jwtAuth: []
      summary: Получение категории
      tags:
      - categories
  /category/update:
    put:
      consumes:
      - application/json
      description: Обновление всех характеристики категории
      parameters:
      - description: данные для обновление категории
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/categoryHandlers.RequestUpdateCategory'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/api.SuccessResponse'
        "400":
          description: Некорректный данные
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Ошибка авторизации
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Категория не найдена
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - jwtAuth: []
      summary: обновление пользователя
      tags:
      - categories
...
securityDefinitions:
  jwtAuth:
    description: 'type "Bearer" после пробел и jwt token, пример: "Bearer zpdgjeawzgp0398tuP29R0J20THVTP9235BHRNr312r346as2..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"